FLBDsim class and methods

A class to simulate the growth of populations aggregated in biomass.

Arguments

...

Empty or FLQuants for 'biomass', 'catch' and 'uncertainty' slots and optionally the values for the rest of the slots.

Value

An object of class FLBDsim.

Details

The FLBDsim has the following slots:

Slots

name

The name of the stock.

desc

A description of the object.

range

The range of the object.

biomass

An FLQuant to store the biomass of the stock.

gB

An FLQuant to store the surplus 4production of the stock.

catch

An FLQuant to store the catch of the stock.

uncertainty

An FLQuant to store the uncertainty that is multiplied to the biomass in every step of the simulation.

covar

An FLQuants to store the covariates that are part of the growth model.

model

A character with the name of the model to simulate the recruitment process.

params

An array with dimension [numb.params, numb.year, numd.season, numb.iteration] with year and season and iteration dependent parameters of the growth model.

alpha

An array with dimension [numb.year, numd.season, numb.iteration] with year, season and iteration dependent value bigger than one which indicates, in percentage, how big can be the biomass in comparison with the carrying capacity.

name

The name of the object.

desc

Character with the description of the object.

range

A numeric vector with the range of the object as in other FLR objects.