FLCatchExt and FLCatchesExt classes and the methods to construct it.

They extend the FLCatch and FLCatches classes defined in FLFleets package. the FLCatch class includes two extra slots alpha and beta used in the Cobb-Douglas production functions.

# S4 method for ANY
FLCatchesExt(object, ...)

# S4 method for missing
FLCatchesExt(object, ...)

# S4 method for list
FLCatchesExt(object)

is.FLCatchesExt(object, ...)

# S4 method for ANY
is.FLCatchesExt(object, ...)

# S4 method for FLQuant
FLCatchExt(object, range = "missing", name = "NA",
  desc = character(0), ...)

# S4 method for missing
FLCatchExt(object, ...)

# S4 method for FLCatchExt,ANY,ANY,ANY
[(x, i, j, k, l, m, n, ..., drop = FALSE)

# S4 method for FLCatchExt,ANY,ANY,FLCatchExt
[(x, i, j, k, l, m, n, ...) <- value

# S4 method for FLCatchExt,FLCatchExt
addFLCatch(e1, e2)

# S4 method for FLCatchExt
catchNames(object)

# S4 method for FLCatchExt,character
catchNames(object) <- value

Arguments

object, x

An object of class FLQuant, missing or FLCatchExt.

...

Other objects to be assigned by name to the class slots

range

Numerical vector with min, max, plusgroup, minyear and maxyear elements as in FLStock object.

name

The name of the stock.

desc

The description of the object.

i, j, k, l, m, n

subindices

drop

logical. Should the dimesions be dropped?

Value

The constructors return an object of class FLCatchExt.

Details

The FLCatchExt object contains a representation of the catch of a fish stock as constructed for the purposes of fleet dynamic modelling. This includes information on removals (i.e. landings and discards), selectivity, weight, price and catch production parameters (catchability and elasticities).

Slots

landings

An FLQuant of dimension [1,num. years, num. units, numb.season, 1, numb. iters] with the total landings in weight of the stock.

landings.n

An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the landings in numbers at age of the stock.

landings.wt

An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the weight at age of the landings.

landings.sel

An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the retention ogive of the metier for this stock. Elements must be between 0 and 1. landings.sel = 1-discards.sel.

discards

An FLQuant of dimension [1,num. years, num. units, numb.season, 1, numb. iters] with the total discards in weight of the stock.

discards.n

An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the discards in numbers at age of the stock.

discards.wt

An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the weight at age of the discards

discards.sel

landings.sel An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the discard ogive of the metier for this stock. Elements must be between 0 and 1. discards.sel. = 1-landings.sel.

catch.q

An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the catchability at age of the stock for the corresponding metier. This is the catchability used in the catch production model.

price

An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the price at age of the stock.

alpha

An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the elsaticity parameter at age of the stock for the corresponding metier. This is one of the parameters used in the catch production model.

beta

An FLQuant of dimension [num. ages ,num. years, num. units, numb.season, 1, numb. iters] with the elsaticity parameter at age of the stock for the corresponding metier. This is one of the parameters used in the catch production model

name

The name of the stock.

desc

A description of the object.

range

The range as in other FLR objects. c("min","max","plusgroup","minyear","maxyear")