FLFleetExt and FLFleetsExt classes and the methods to construct it.

They extend the FLFleetExt and FLFleetsExt classes defined in FLFleets package. The only different is that that the metiers slot is a FLMetiersExt object

# S4 method for ANY
FLFleetsExt(object, ...)

# S4 method for missing
FLFleetsExt(object, ...)

# S4 method for list
FLFleetsExt(object)

is.FLFleetsExt(object, ...)

# S4 method for ANY
is.FLFleetsExt(object, ...)

# S4 method for FLMetiersExt
FLFleetExt(object, ...)

# S4 method for FLMetierExt
FLFleetExt(object, ...)

# S4 method for FLCatchesExt
FLFleetExt(object, ...)

# S4 method for FLCatchExt
FLFleetExt(object, ...)

# S4 method for FLFleetExt
FLFleetExt(object, metier, catch, ...)

# S4 method for missing
FLFleetExt(object, ...)

# S4 method for FLFleetExt,ANY,missing,ANY
[(x, i, drop = FALSE)

# S4 method for FLFleetExt,ANY,ANY,ANY
[(x, i, j, drop = FALSE)

# S4 method for FLFleetExt,ANY,missing
[[(x, i, drop = FALSE)

Arguments

object

An object of class FLQuant, missing or FLMetierExt.

...

Other objects to be assigned by name to the class slots

metier

A name of one of the elements in FLMetiers object.

catch

A name of one of the elements in FLCatches object.

range

Numerical vector with min, max, plusgroup, minyear and maxyear elements as in FLStock object.

name

The name of the fleet.

desc

The description of the object.

metiers

An object of class FLMetierExt or FLMetiersExt.

Value

The constructors return an object of class FLFleetExt.

Details

The FLFleetExt object contains a representation of a fishing fleet as constructed for the purposes of fleet dynamic modelling. This includes information on effort, fixed-cost, capacity, crew-share, metiers and variable costs.