FLMetierExt and FLMetiersExt classes and the methods to construct it.

They extend the FLMetier and FLMetiers classes defined in FLFleets package. The only different is that that the catches slot is a FLCatchesExt object

# S4 method for ANY
FLMetiersExt(object, ...)

# S4 method for missing
FLMetiersExt(object, ...)

# S4 method for list
FLMetiersExt(object)

is.FLMetiersExt(object, ...)

# S4 method for ANY
is.FLMetiersExt(object, ...)

# S4 method for FLCatchExt
FLMetierExt(catches, gear = "NA", ...)

# S4 method for FLCatchesExt
FLMetierExt(catches, gear = "NA", ...)

# S4 method for FLQuant
FLMetierExt(catches, gear = "NA", ...)

# S4 method for missing
FLMetierExt(catches, gear = "NA", ...)

# S4 method for FLMetierExt,ANY,missing,ANY
[(x, i, drop = FALSE)

# S4 method for FLMetierExt,ANY,missing
[[(x, i, drop = FALSE)

Arguments

object

An object of class FLQuant, missing or FLCatchExt.

...

Other objects to be assigned by name to the class slots

catches

An object of class FLCatchesExt.

gear

A character vector with the name of the gear used in the metier.

range

Numerical vector with min, max, plusgroup, minyear and maxyear elements as in FLStock object.

name

The name of the metier.

desc

The description of the object.

Value

The constructors return an object of class FLMetierExt.

Details

The FLMetierExt object contains a representation of the metier of a fishing fleet as constructed for the purposes of fleet dynamic modelling. This includes information on effortshare and variable costs.