FLSRsim class

This class is used to store the necessary information to simulate the recruitment process within FLBEIA.

Arguments

...

Any of the slots in FLSRsim class.

Value

An object of class FLSRsim.

Details

The FLSRsim class contains a representation of the recruitment process of a fish stock. This includes information on recruitment, ssb, recruitment model, uncertainty and distribution of the recruitment along seasons. The slots in this class:

Slots

rec

An FLQuant with to store the recruitment.

ssb

An FLQuant with to store the ssb.

covar

An FLQuants to store the covariates considered in the stock-recruitment model, one FQuant per covariate.

uncertainty

An FLQuant with to store the uncertainty that is multiplied to the recruitment point estimate in each step of the simulation.

proportion

An FLQuant with values between 0 and 1 to indicate how the recruitment in each of the time steps is distributed along season.

model

A character with the name of the model to simulate the recruitment process.

params

An array with dimensions [number of params,number of years,number of seasons,number of iteration].

timelag

A matrix [2, number of seasons]. The element (1,j) indicates the time lag between the spawning and recruitment year and the element (2,j) the season in which the recruitment was spawn.

name

A character with the name of the stock.

desc

A description of the object.

range

A numeric vector with c(min, max,plusgroup, minyear, maxyear) as in the rest of the FLR objects.