Biological summary functions

These functions return the biomass (B), fishing mortality (F), spawning stock biomass (SSB), recruitment (R), catches (C), landings (L) and discards (D) indicators.

F_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

SSB_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

B_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

R_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

C_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

L_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

D_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

summary_flbeia(obj, years = dimnames(obj$biols[[1]]@n)$year)

Arguments

obj

The output of the FLBEIA function.

years

The years for which the indicators are extracted.

Value

B_flbeia, F_flbeia... return an array with three dimensions (stock, year and iter). The summary_flbeia function returns an array with 4 dimensions (stock, year, iter, indicator) with the value of all the indicators.

Details

B_flbeia this function computes SSB.

F_flbeia ithis function computes fishing mortiality.

SSB_flbeia this function computes spawning stock biomass by species.

R_flbeia this function computes recruitment by stock. If the stock is defined by age this function the recruiment is computed. ; If the stock is follows a biomass dynamics, this function gives the growth.

C_flbeia this function computes catches by fleets and stock.

L_flbeia this function computes landings by fleets and stock.

D_flbeia ithis function computes the discards by fleets and stock.