Harvest Control Rules (HCRs)

There are several HCRs available in FLBEIA which are used within the main function FLBEIA to generate the management advice in each step. But they can also be used independently.

annualTAC(stocks, advice, advice.ctrl, year, stknm, ...)

aneHCRE(stocks, advice, advice.ctrl, year, stknm, ...)

F2CatchHCR(stocks, advice, advice.ctrl, year, stknm, ...)

FroeseHCR(stocks, advice, advice.ctrl, year, stknm, ...)

IcesHCR(stocks, advice, advice.ctrl, year, stknm, ...)

neaMAC_ltmp(stocks, advice, advice.ctrl, year, stknm, ...)

annexIVHCR(indices, advice, advice.ctrl, year, stknm, ...)

ghlHCR(indices, advice, advice.ctrl, year, stknm, ...)

MAPHCR(stocks, advice, advice.ctrl, year, stknm, ...)

CFPMSYHCR(stocks, advice, advice.ctrl, year, stknm, ...)

MultiStockHCR(stocks, indices, advice, advice.ctrl, year, stknm, ...)

pidHCR(indices, advice, advice.ctrl, year, stknm, ...)

pidHCRItarg(indices, advice, advice.ctrl, year, stknm, ...)

little2011HCR(indices, advice, advice.ctrl, year, stknm, ...)

Arguments

stocks

And FLStocks object.

advice

A list with two FLQuant elements, TAC and quota.share. TAC is an FLQuant with quant dimension equal to the number of stocks in biols object, the names used in in the quant dimension must be equal to those used in biols. quota.share is a list with one element per stock in biols object indicating the quota share per stock and fleet. The quant dimension of the elements must be equal to the number of fleets and the names used must be equal to those in fleets objects.

advice.ctrl

A list with the settings to control the advice model for each stock (the HCR for each stock, the reference points used in the HCR, additional parameters, ...)

year

The position of the assessment year in the stocks and advice objects.

stknm

The name of the stock for which advice is being generated.

...

Any extra arguments needed for specific HCRs.

indices

A list of FLIndices. Each element must correspond with one of the stocks in biols object.

Value

The advice input object updated with the management advice (TAC) generated by the HCR.

Details

There are two types of HCRs model-free HCRs and model-Based HCRs. Model-free HCRs use abundance indices to generate the advice and hence it use FLIndices object as input data. Model-based HCRs use estimates of stock abundance and stock exploitation level to generate the advice.

Model-Free HCRs: annexIVHCR, ghlHCR, little2011HCR, pidHCR and pidHCRItarg.

Model-Based HCRs: aneHCRE, annualTAC, CFPMSYHCR, F2CatchHCR, FroeseHCR, IcesHCR, MAPHCR, neaMAC_ltmp.

aneHCRE: The HCR used in the bay of biscay anchovy long term management plan.

annexIVHCR: The HCR used by EC and ICES to generate the TAC advice for data poor stocks.

annualTAC: A HCR that generates annual TAC advice. The HCR provides the whole flexibility of fwd.

CFPMSYHCR: HCR adapting the MAP HCR to allow flexibility in the year Fmsy is achieved. The user can specify the year in which you aim to reach Fmsy, with a linear transition between Fsq to Fmsy in the intervening years

F2CatchHCR: This HCR transforms the fishing mortality advice given as input data to catch advice without any other restriction.

FroeseHCR: The HCR defined in the paper by Froese, Branch et al. in Fish and Fisheries 2010.

ghlHCR: The model-free HCR used in the management of greenland-halibut

IcesHCR: The HCR used by ICES to generate TAC advice in the MSY framework.

little2011HCR: The HCR defined in the paper by Little et al. in ICES Journal of Marine Science 2011.

MAPHRC: The HCR proposed by the EC in the evaluation on multi-annual management plans in 2015.

MultiStockHRC: A HCR that produces TAC advice for several stocks simultaneously. It uses a fishing mortality target and an upper bound to conciliate the TAC advices. In the case of stocks without exploitation rate estimates it uses the catch.

neaMAC_ltmp: The HCR used in the north-east atlantic mackerel long term management plan. It is a particular case of the IcesHCR.

pidHCR, pidHCRItarg: The HCRs defined in the paper by Pomaerede et al. in Aquatic Living Resources 2010.

The HCRs are documented in detail in the manual of the library.

Examples

## Not run: ------------------------------------
# library(FLBEIA)
# library(FLAssess)          # required to use the IcesHCR. Not available for win64
# library(FLash)             # required to use the IcesHCR. Not available for win64
# library(ggplot2)  
# 
# # Load the data to run FLBEIA in a one stock one fleet example using the HCR used by ICES in the MSY framework. 
#  data(one) 
#  
# oneAdv$TAC[,ac(2009:2025)] <- NA # Put NA-s in the projection years to check how the function fills the advice object.
#  
# res <- IcesHCR(oneSt, oneAdv, oneAdvC, 19, 'stk1') # The value printed in the screen is the fishing mortality used in the advice.
#  
# res$TAC[,'2009']    # The resulting management advice.
## ---------------------------------------------