Summary of the FLBEIA output

Summarize the results of the simulation in data frames.

bioSum(obj, stknms = "all", years = dimnames(obj$biols[[1]]@n)$year,
  long = TRUE, scenario = "bc")

bioSumQ(obj, prob = c(0.95, 0.5, 0.05))

fltSum(obj, flnms = "all", years = dimnames(obj$biols[[1]]@n)$year,
  byyear = TRUE, long = TRUE, scenario = "bc")

fltSumQ(obj, prob = c(0.95, 0.5, 0.05))

fltStkSum(obj, flnms = names(obj$fleets), stknms = catchNames(obj$fleets),
  years = dimnames(obj$biols[[1]]@n)[[2]], byyear = TRUE, long = TRUE,
  scenario = "bc")

fltStkSumQ(obj, prob = c(0.95, 0.5, 0.05))

mtStkSum(obj, flnms = names(obj$fleets), stknms = catchNames(obj$fleets),
  years = dimnames(obj$biols[[1]]@n)[[2]], byyear = TRUE, long = TRUE,
  scenario = "bc")

mtStkSumQ(obj, prob = c(0.95, 0.5, 0.05))

mtSum(obj, flnms = names(obj$fleets),
  years = dimnames(obj$biols[[1]]@n)[[2]], byyear = TRUE, long = TRUE,
  scenario = "bc")

mtSumQ(obj, prob = c(0.95, 0.5, 0.05))

advSum(obj, stknms = catchNames(obj$fleets),
  years = dimnames(obj$biols[[1]]@n)[[2]], long = TRUE, scenario = "bc")

advSumQ(obj, prob = c(0.95, 0.5, 0.05))

riskSum(obj, stknms = names(obj$biols), Bpa, Blim, Prflim,
  flnms = names(obj$fleets), years = dimnames(obj$biols[[1]]@n)[[2]],
  scenario = "bc")

npv(obj, discF = 0.05, y0, flnms = names(obj$fleets),
  years = dimnames(obj$biols[[1]]@n)[[2]], scenario = "bc")

npvQ(obj, prob = c(0.05, 0.5, 0.95))

vesselSum(obj, flnms = "all", years = dimnames(obj$biols[[1]]@n)$year,
  byyear = TRUE, long = TRUE, scenario = "bc")

vesselSumQ(obj, prob = c(0.95, 0.5, 0.05))

vesselStkSum(obj, flnms = names(obj$fleets),
  stknms = catchNames(obj$fleets), years = dimnames(obj$biols[[1]]@n)[[2]],
  byyear = TRUE, long = TRUE, scenario = "bc")

vesselStkSumQ(obj, prob = c(0.95, 0.5, 0.05))

Arguments

obj

The output of the FLBEIA function.

stknms

Names of the stock for which the indicators will be calculated.

years

the names of the years for which the indicators will be calculated.

long

logical. The data frame should be constructed using long or wide format? Default TRUE.

scenario

a character string with the name of the scenario corresponding with obj. Default bc.

prob

a numeric vector with the probabilities used to calculate the quantiles.

flnms

Names of the fleet for which the indicators will be calculated.

byyear

logical. The indicators should be provided at season or year level? Default TRUE.

Bpa

named numeric vector with one element per stock in stknms. The precautionary approach stock spawning biomass used in riskSum function to calculate biological risk yearly.

Blim

named numeric vector with one element per stock in stknms. The limit stock spawning biomass used in riskSum function to calculate biological risk yearly.

Prflim

named numeric vector with one element per fleet in flnms. The limit profit level used in riskSum function to calculate biological risk yearly.

Value

The data frames can be of wide or long format. In long format all the indicators are in the same column. There is one column, indicator, for the name of the indicator and a second one value for the numeric value of the indicator. In the wide format each of the indicators correspond with one column in the data frame. The long format it is recommendable to work with ggplot2 functions for example while the wide format it is more efficient for memory allocation and speed of computations.

The quantile version of the summaries, fooQ, returns the quantiles of the indicators. In the long format as many columns as elements in prob are created. The name of the columns are the elements in prob preceded by a q. In the wide format for each of the indicators as many columns as elements in prob are created. The names of the colums are the elements in prob preceded by q_name_of_the_indicator.

Details

advSum, advSumQ: Data frame with the indicators related with the management advice (TAC). The indicators are: "catch", "discards", "discRat", "landings", "quotaUpt" and "tac".

bioSum, bioSumQ: Data frame with the biological indicators. The indicators are: "biomass", "catch", "catch.iyv", "discards", "disc.iyv", "f", "landings", "land.iyv", "rec" and "ssb".

fltSum, fltSumQ: Data frame with the indicators at fleet level. The indicators are: "capacity", "catch", "costs", "discards", "discRat", "effort", "fcosts", "gva", "income", "landings", "netProfit", "nVessels", "price", "profits", "quotaUpt", "salaries", "vcosts" and "profitability".

fltStkSum, fltStkSumQ: Data frame with the indicators at fleet and stock level. The indicators are: "landings", "discards", "catch", "price", "quotaUpt", "tacshare", "discRat" and "quota".

npv: A data frame with the net present value per fleet over the selected range of years.

mtSum, mtSumQ: Data frame with the indicators at fleet. The indicators are: "effshare", "effort", "income" and "vcost".

mtStkSum, mtStkSumQ: Data frame with the indicators at fleet and metier level. The indicators are: "catch", "discards", "discRat", "landings" and "price".

riskSum: A data frame with the risk indicators. The indicators are: "pBlim", "pBpa" and "pPrflim".

vesselSum, vesselSumQ: Data frame with the indicators at vessel level. The indicators are: "catch", "costs", "discards", "discRat", "effort", "fcosts", "gva", "income", "landings", "netProfit", "price", "profits", "quotaUpt", "salaries", "vcosts" and "profitability".

vesselStkSum, vesselStkSumQ: Data frame with the indicators at vessel and stock level. The indicators are: "landings", "discards", "catch", "price", "quotaUpt", "tacshare", "discRat" and "quota".

summary_flbeia: An array with four dimensions: stock, year, iteration, indicator. The indicators are: recruitment, ssb, f, biomass, catch, landings and discards.

ecoSum_damara: ecoSum built in the framework of Damara project.

The data frames

Examples

## Not run: ------------------------------------
# 
# library(FLBEIA)
# 
# # Apply the summary functions to the examples runs in FLBEIA help page.
# # Test the different arguments in summary function.
# 
# data(res_flbeia)
# #------------------------------------------------
# # Example One: One stock, one fleet, one iter.
# #------------------------------------------------
# s0_bio    <- bioSum(s0)
# s0_flt    <- fltSum(s0)
# s0_fltStk <- fltStkSum(s0)
# s0_mt     <- mtSum(s0)
# s0_mtStk  <- mtStkSum(s0)
# s0_adv    <- advSum(s0)
# 
# head(s0_bio)
# head(s0_flt)
# head(s0_fltStk)
# head(s0_mt)
# head(s0_mtStk)
# head(s0_adv)
# 
# s0_bioQ    <- bioSumQ(s0_bio)
# s0_fltQ    <- fltSumQ(s0_flt)
# s0_fltStkQ <- fltStkSumQ(s0_fltStk)
# s0_mtQ     <- mtSumQ(s0_mt)
# s0_mtStkQ  <- mtStkSumQ(s0_mtStk)
# s0_advQ    <- advSumQ(s0_adv)
# 
# head(s0_bioQ)
# head(s0_fltQ)
# head(s0_fltStkQ)
# head(s0_mtQ)
# head(s0_mtStkQ)
# head(s0_advQ)
# 
# # Wide format
# s0_bio    <- bioSum(s0, long = FALSE, years = ac(2016:2020))
# s0_flt    <- fltSum(s0, long = FALSE, years = ac(2016:2020))
# s0_fltStk <- fltStkSum(s0, long = FALSE, years = ac(2016:2020))
# s0_mt     <- mtSum(s0, long = FALSE, years = ac(2016:2020))
# s0_mtStk  <- mtStkSum(s0, long = FALSE, years = ac(2016:2020))
# s0_adv    <- advSum(s0, long = FALSE, years = ac(2016:2020))
# 
# head(s0_bio)
# head(s0_flt)
# head(s0_fltStk)
# head(s0_mt)
# head(s0_mtStk)
# head(s0_adv)
# 
# s0_bioQ    <- bioSumQ(s0_bio)
# s0_fltQ    <- fltSumQ(s0_flt)
# s0_fltStkQ <- fltStkSumQ(s0_fltStk)
# s0_mtQ     <- mtSumQ(s0_mt)
# s0_mtStkQ  <- mtStkSumQ(s0_mtStk)
# s0_advQ    <- advSumQ(s0_adv)
# 
# head(s0_bio)
# head(s0_flt)
# head(s0_fltStk)
# head(s0_mt)
# head(s0_mtStk)
# head(s0_adv)
# 
# # Wide format with seasonal disaggregation. No seasonal disagregation available for bio and adv summaries.
# 
# s0_bio    <- bioSum(s0, long = FALSE) # Biol summary is only by year.
# s0_flt    <- fltSum(s0, long = FALSE, byyear = FALSE)
# s0_fltStk <- fltStkSum(s0, long = FALSE, byyear = FALSE)
# s0_mt     <- mtSum(s0, long = FALSE, byyear = FALSE)
# s0_mtStk  <- mtStkSum(s0, long = FALSE, byyear = FALSE)
# s0_adv    <- advSum(s0, long = FALSE) # Advice summary is only by year.
# 
# s0_bioQ    <- bioSumQ(s0_bio)
# s0_fltQ    <- fltSumQ(s0_flt)
# s0_fltStkQ <- fltStkSumQ(s0_fltStk)
# s0_mtQ     <- mtSumQ(s0_mt)
# s0_mtStkQ  <- mtStkSumQ(s0_mtStk)
# s0_advQ    <- advSumQ(s0_adv)
# s0_bio    <- bioSum(s0, long = TRUE) # Biol summary is only by year.
# s0_flt    <- fltSum(s0, long = TRUE, byyear = FALSE)
# s0_fltStk <- fltStkSum(s0, long = TRUE, byyear = FALSE)
# s0_mt     <- mtSum(s0, long = TRUE, byyear = FALSE)
# s0_mtStk  <- mtStkSum(s0, long = TRUE, byyear = FALSE)
# s0_adv    <- advSum(s0, long = TRUE) # Advice summary is only by year.
# 
# s0_bioQ    <- bioSumQ(s0_bio)
# s0_fltQ    <- fltSumQ(s0_flt)
# s0_fltStkQ <- fltStkSumQ(s0_fltStk)
# s0_mtQ     <- mtSumQ(s0_mt)
# s0_mtStkQ  <- mtStkSumQ(s0_mtStk)
# s0_advQ    <- advSumQ(s0_adv)
# 
# 
# #------------------------------------------------
# # Example OneIters: As one but with iterations.
# #------------------------------------------------
# s1_bio    <- bioSum(s1, scenario = 'with_iters')
# s1_flt    <- fltSum(s1, scenario = 'with_iters')
# s1_fltStk <- fltStkSum(s1, scenario = 'with_iters')
# s1_mt     <- mtSum(s1, scenario = 'with_iters')
# s1_mtStk  <- mtStkSum(s1, scenario = 'with_iters')
# s1_adv    <- advSum(s1, scenario = 'with_iters')
# 
# s1_bioQ    <- bioSumQ(s1_bio)
# s1_fltQ    <- fltSumQ(s1_flt)
# s1_fltStkQ <- fltStkSumQ(s1_fltStk)
# s1_mtQ     <- mtSumQ(s1_mt)
# s1_mtStkQ  <- mtStkSumQ(s1_mtStk)
# s1_advQ    <- advSumQ(s1_adv)
# 
# s1_bio    <- bioSum(s1, long = FALSE, years = ac(2016:2020))
# s1_flt    <- fltSum(s1, long = FALSE, years = ac(2016:2020))
# s1_fltStk <- fltStkSum(s1, long = FALSE, years = ac(2016:2020))
# s1_mt     <- mtSum(s1, long = FALSE, years = ac(2016:2020))
# s1_mtStk  <- mtStkSum(s1, long = FALSE, years = ac(2016:2020))
# s1_adv    <- advSum(s1, long = FALSE, years = ac(2016:2020))
# 
# 
# s1_bioQ    <- bioSumQ(s1_bio)
# s1_fltQ    <- fltSumQ(s1_flt)
# s1_fltStkQ <- fltStkSumQ(s1_fltStk)
# s1_mtQ     <- mtSumQ(s1_mt)
# s1_mtStkQ  <- mtStkSumQ(s1_mtStk)
# s1_advQ    <- advSumQ(s1_adv)
# 
# 
# s1_bio    <- bioSum(s1, long = FALSE) # Biol summary is only by year.
# s1_flt    <- fltSum(s1, long = FALSE, byyear = FALSE)
# s1_fltStk <- fltStkSum(s1, long = FALSE, byyear = FALSE)
# s1_mt     <- mtSum(s1, long = FALSE, byyear = FALSE)
# s1_mtStk  <- mtStkSum(s1, long = FALSE, byyear = FALSE)
# s1_adv    <- advSum(s1, long = FALSE) # Advice summary is only by year.
# 
# s1_bioQ    <- bioSumQ(s1_bio)
# s1_fltQ    <- fltSumQ(s1_flt)
# s1_fltStkQ <- fltStkSumQ(s1_fltStk)
# s1_mtQ     <- mtSumQ(s1_mt)
# s1_mtStkQ  <- mtStkSumQ(s1_mtStk)
# s1_advQ    <- advSumQ(s1_adv)
# 
# 
# s1_bio    <- bioSum(s1, long = TRUE) # Biol summary is only by year.
# s1_flt    <- fltSum(s1, long = TRUE, byyear = FALSE)
# s1_fltStk <- fltStkSum(s1, long = TRUE, byyear = FALSE)
# s1_mt     <- mtSum(s1, long = TRUE, byyear = FALSE)
# s1_mtStk  <- mtStkSum(s1, long = TRUE, byyear = FALSE)
# s1_adv    <- advSum(s1, long = TRUE) # Advice summary is only by year.
# 
# s1_bioQ    <- bioSumQ(s1_bio)
# s1_fltQ    <- fltSumQ(s1_flt)
# s1_fltStkQ <- fltStkSumQ(s1_fltStk)
# s1_mtQ     <- mtSumQ(s1_mt)
# s1_mtStkQ  <- mtStkSumQ(s1_mtStk)
# s1_advQ    <- advSumQ(s1_adv)
# 
# s1_risk <- riskSum(s1, Bpa = c(stk1= 900), Blim = c(stk1 = 600), Prflim = c(flt1 = 0), scenario = 'alternative')
# 
# s1_npv  <- npv(s1, y0 = '2014')
# 
# #------------------------------------------------
# # Example Multi: Two stock, two fleet, four iters.
# #------------------------------------------------
# s2_bio    <- bioSum(s2)
# s2_flt    <- fltSum(s2)
# s2_fltStk <- fltStkSum(s2)
# s2_mt     <- mtSum(s2)
# s2_mtStk  <- mtStkSum(s2)
# s2_adv    <- advSum(s2)
# 
# s2_bioQ    <- bioSumQ(s2_bio)
# s2_fltQ    <- fltSumQ(s2_flt)
# s2_fltStkQ <- fltStkSumQ(s2_fltStk)
# s2_mtQ     <- mtSumQ(s2_mt)
# s2_mtStkQ  <- mtStkSumQ(s2_mtStk)
# s2_advQ    <- advSumQ(s2_adv)
# 
# s2_bio    <- bioSum(s2, long = FALSE, years = ac(2016:2020))
# s2_flt    <- fltSum(s2, long = FALSE, years = ac(2016:2020))
# s2_fltStk <- fltStkSum(s2, long = FALSE, years = ac(2016:2020))
# s2_mt     <- mtSum(s2, long = FALSE, years = ac(2016:2020))
# s2_mtStk  <- mtStkSum(s2, long = FALSE, years = ac(2016:2020))
# s2_adv    <- advSum(s2, long = FALSE, years = ac(2016:2020))
# 
# 
# s2_bioQ    <- bioSumQ(s2_bio)
# s2_fltQ    <- fltSumQ(s2_flt)
# s2_fltStkQ <- fltStkSumQ(s2_fltStk)
# s2_mtQ     <- mtSumQ(s2_mt)
# s2_mtStkQ  <- mtStkSumQ(s2_mtStk)
# s2_advQ    <- advSumQ(s2_adv)
# 
# 
# s2_bio    <- bioSum(s2, long = FALSE) # Biol summary is only by year.
# s2_flt    <- fltSum(s2, long = FALSE, byyear = FALSE)
# s2_fltStk <- fltStkSum(s2, long = FALSE, byyear = FALSE)
# s2_mt     <- mtSum(s2, long = FALSE, byyear = FALSE)
# s2_mtStk  <- mtStkSum(s2, long = FALSE, byyear = FALSE)
# s2_adv    <- advSum(s2, long = FALSE) # Advice summary is only by year.
# 
# s2_bioQ    <- bioSumQ(s2_bio)
# s2_fltQ    <- fltSumQ(s2_flt)
# s2_fltStkQ <- fltStkSumQ(s2_fltStk)
# s2_mtQ     <- mtSumQ(s2_mt)
# s2_mtStkQ  <- mtStkSumQ(s2_mtStk)
# s2_advQ    <- advSumQ(s2_adv)
# 
# 
# s2_bio    <- bioSum(s2, long = TRUE) # Biol summary is only by year.
# s2_flt    <- fltSum(s2, long = TRUE, byyear = FALSE)
# s2_fltStk <- fltStkSum(s2, long = TRUE, byyear = FALSE)
# s2_mt     <- mtSum(s2, long = TRUE, byyear = FALSE)
# s2_mtStk  <- mtStkSum(s2, long = TRUE, byyear = FALSE)
# s2_adv    <- advSum(s2, long = TRUE) # Advice summary is only by year.
# 
# s2_bioQ    <- bioSumQ(s2_bio)
# s2_fltQ    <- fltSumQ(s2_flt)
# s2_fltStkQ <- fltStkSumQ(s2_fltStk)
# s2_mtQ     <- mtSumQ(s2_mt)
# s2_mtStkQ  <- mtStkSumQ(s2_mtStk)
# s2_advQ    <- advSumQ(s2_adv)
# 
# s2_npv  <- npv(s2, y0 = '2014')
# risk_s2 <- riskSum(s2, Bpa = c(stk1= 135000, stk2 = 124000), Blim = c(stk1= 96000, stk2 = 89000), Prflim = c(flt1 = 0, flt2 = 0), scenario = 'alternative')
# 
## ---------------------------------------------