FLBEIA easy conditioning: biols argument creator

create.biols.data function creates an FLBiols object.

create.biols.data(yrs, ns, ni, stks.data)

Arguments

yrs

A vector with c(first.yr,proj.yr, last.yr) where

first.yr: First year of simulation (number).

proj.yr: First year of projection (number).

last.yr: Last year of projection (number).

ns

Number of seasons (number).

ni

Number of iterations (number).

stks.data

A list with the name of the stks and the following elements:

stk.unit: Number of units of the stock (number).

stk.age.min: Minimum age class of the stock (number).

stk.age.max: Maximum age class of the stock (number).

stk_n.flq: Numbers at age in the population(FLQuant).

stk_wt.flq: Weight at age of an individual (FLQuant).

stk_m.flq: Mortality rate at age of the population (FLQuant).

stk_fec.flq: Fecundity at age (FLQuant).

stk_mat.flq: Percentage of mature individuals at age (FLQuant).

stk_spwn.flq: Proportion of time step at which spawning ocurrs (FLQuant).

stk.range.plusgroup: Plusgroup age (number).

stk.range.minyear: Minimum year (number).

stk.range.maxyear: Maximum year (number).

stk_range.minfbar: Minimum age to calculate average fishing mortality (number).

stk_range.maxfbar: Maximum age to calculate average fishing mortality (number).

stk_biol.proj.avg.yrs: Historic years to calculate the average of spwn, fec, m and wt for the projection (vector).

Value

An FLBiol object