fleets.ctrl object creator

It creates the fleets.ctrl object to be used in the call to the main function FLBEIA.

create.fleets.ctrl(fls, n.fls.stks, fls.stksnames, catch.threshold = NULL,
  seasonal.share = NULL, effort.models = NULL, capital.models = NULL,
  catch.models = NULL, price.models = NULL, flq, ...)

Arguments

fls

character vector with fleet names

n.fls.stks

numeric vector with the same length as fls with the declaration of the number of stocks caugth by each of the fleets.

fls.stksnames

character vector with length = sum(n.fls.stks), with the names of the stocks caught by the fleet, the vector must follow the order used in the previous argument.

the first n.fls.stks[1] elements correspond to the stocks caught by the first fleet in fls

the following n.fls.stks[2] elements correspond to the stocks caught by the second fleet in fls

and so on.

catch.threshold

if(NULL) => 0.9 for all the stocks (NULL is the default) else it must be an FLQuant with dim = c(nstks,ny,1,ns,nit)

seasonal.share

an FLQuant with dimension [num. fleets, num. years, 1,num. seasons, 1, num. iterations] with elements between 0 and 1 to indicate how the quota of each fleet is distributed along seasons. The sum along seasons (seasonSums) must return an FLQuant with all elements equal to 1.

effort.models

characted vector with the same length as fls with the effort model followed by each of the fleet. the first element correspond with the effort model of the first fleet in fls, the second with the second and so on. The default is NULL in which case 'fixedEffort' is used for **all** the fleets.

capital.models

characted vector with the same length as fls with the capital model followed by each of the fleet. the first element correspond with the capital model of the first fleet in fls, the second with the second and so on. The default is NULL in which case 'fixedCapital' is used for **all** the fleets.

catch.models

characted vector with the same length as sum(n.fls.stks) with the catch model followed by each of the fleet for each stock. the first element correspond with the catch model of the first fleet in fls and the first stock in fls.stksnames, the second with the second and so on. The default is NULL in which case 'CobbDouglasAge' is used for **all** the fleets.

price.models

characted vector with the same length as sum(n.fls.stks) with the price model followed by each of the fleet for each stock. the first element correspond with the price model of the first fleet in fls and the first stock in fls.stksnames, the second with the second and so on. The default is NULL in which case 'fixedPrice' is used for **all** the fleets.

flq

An flquant to give structure to the FLQuants to be used within the function, the dimensioan and dimnames in 'year', 'season' and 'iter' will be used to create the necessary FLQuants.

...

Any extra arguments necessary in the model specific creators. '...' are extracted using 'list(...)', this generates a named list with the extra arguments. To assure the correct functioning the extra arguments must have a name, for example, elas = FLQuant(1,dimnames = DimsNms).

Value

A list of lists with the basic structure of the fleets.ctrl object.