FLBEIA easy conditioning: indices argument creator

create.indices.data function creates an FLIndices object

create.indices.data(yrs, ns, ni, stks.data)

Arguments

yrs

A vector with c(first.yr,proj.yr, last.yr) where

first.yr: First year of simulation (number).

proj.yr: First year of projection (number).

last.yr: Last year of projection (number).

ns

Number of seasons (number).

ni

Number of iterations (number).

stks.data

A list with the names of the stocks with indices and the following elements:

stk.unit: Number of units of the stock (number).

stk.age.min: Minimum age class of the stock (number).

stk.age.max: Maximum age class of the stock (number).

stk_indices: Name of indices for the stock 'stk' (vector).

stk_ind_index.flq: Historical index data for index 'ind' of stock 'stk' (FLQuant).

Optionals:

stk_ind_type: Type of index (character).

stk_ind_distribution: Name of the statistical distribution of the 'ind' index values for stock 'stk' (character).

stk_ind_index.var.flq: Variability in 'ind' index of stock 'stk' (FLQuant).

stk_ind_index.q.flq: Catchability for 'ind' index of stock 'stk' (FLQuant).

stk_ind_catch.n.flq: Catch at age in numbers for 'ind' index of stock 'stk'(FLQuant).

stk_ind_catch.wt.flq: Mean weight at age in the catch for 'ind' index of stock 'stk' (FLQuant).

stk_ind_effort.flq: Effort for 'ind' index of stock 'stk' (FLQuant).

stk_ind_sel.pattern.flq: Selection pattern for 'ind' index of stock 'stk' (FLQuant).

stk_ind_range.min: Minimum age in 'ind' index of stock 'stk' (number).

stk_ind_range.max: Maximum age in 'ind' index of stock 'stk' (number).

stk_ind_range.plusgroup: Plusgroup age in 'ind' index of stock 'stk' (number).

stk_ind_range.minyear: First year with 'ind' index data of stock 'stk' (number).

stk_ind_range.maxyear: Last year with 'ind' index data of stock 'stk' (number).

stk_ind_range.startf: Minimum age for calculating average fishing mortality for 'ind' index of stock 'stk' (number).

stk_ind_range.endf: Maximum age for calculating average fishing mortality for 'ind' index of stock 'stk' (number).

Value

An FLIndices object.