obs.ctrl object creator

It creates the obs.ctrl object to be used in the call to the main function FLBEIA.

create.obs.ctrl(stksnames, n.stks.inds = NULL, stks.indsnames = NULL,
  stkObs.models = NULL, indObs.models = NULL, immediate = FALSE, ...)

Arguments

stksnames
A vector with the name of the stocks in the OM.

n.stks.inds
numeric vector with the same length as stksnames with the declaration of the number of abundance indices per stock, the order must be the same used in stksnames.

stks.indsnames
NULL or a character vector of length equal to sum(n.stks.inds) with the names of the indices per stock. The order must be the same used in the previous arguments.

stkObs.models
A character vector of the same length as stksnames with the name of the model used to observed stock data.

indObs.models
A character vector of the same length as stks.indsnames with the name of the model used to generate the abundance indices.

immediate
logical, indicating if the warnings should be output immediately.

...
any extra arguments necessary in the model specific creators. '...' are extracted using 'list(...)', this generates a named list with the extra arguments. To assure the correct functioning the extra arguments must have a name.

Value

A list of lists with the basic structure of the obs.ctrl object.