FLBEIA datasets

Example datasets for the classes defined in FLBEIA.

Details

one: A dataset for running FLBEIA. Example with one stock (age-structured), one fleet, annual steps (one season) and one iteration.

oneBio (FLBiols) : Biological information on the stock. In this case, the stock is age-structured.

oneSR (FLSRsim) : Stock-recruitment model for the stock in oneBio object.

oneFl (FLFleetsExt) : Information on the fleet and the metier considered.

oneCv (list of FLQuants) : Covariates information. In this case all are economic indicators.

oneIndAge and oneIndBio (list of FLIndices) : Indices, if avalable, for the different stocks in oneBio. Where oneIndAge and oneIndBio are indices with estimates in numbers at age and total biomass, respectively.

oneAdv (list) : Information on TAC and quota share.

oneMainC (list) : Settings to control the main function FLBEIA. The simulation years.

oneBioC (list) : Settings to control the biological operating model for the stock in oneBio object.

oneFlC (list) : Settings to control the fleet operating model for the fleet in oneFl object.

oneCvC (list) : Settings to control the covar operating model for each covariate in covars object. In this case all the covariates are fixed.

oneObsC, oneObsCIndAge and oneObsCIndBio (list) : Settings to control the observation model for the stock in oneBio object. In oneObsC the stock is observed without error and there are no indices available. Alternative control settings are available for cases when indices are observed, oneObsCIndAge and oneObsCIndBio.

oneAssC (list) : Settings to control the assessment model for each stock in oneBio object. In this case, no assessment is carried out.

oneAdvC (list) : Settings to control the advice model for the stock in oneBio object.

oneIt: A dataset for running FLBEIA. Same as one dataset, but with three iterations.

multi: A dataset for running FLBEIA. Example with two stocks (one age-structured and the other in biomass), two fleets (with 2 metiers each), four seasons and one iteration.

multiBio (FLBiols) : Biological information on the stocks (one is age-structured and the other one in total biomass).

multiSR (FLSRsim) : Stock-recruitment models for the age-structured stock in multiBio object. In this case a Beverton-Holt (bevholt) is selected.

multiBD (FLSRsim) : Biomass dynamic model for the stock in biomass in multiBio object. In this case Pella-Tomlinson model (PellaTom) is selected.

multiFlC (FLFleetsExt) : Information on the fleets and metiers. In this case there are two fleets, each one with two metiers, all of them capturing both stocks in multiBio object.

multiCv (list of FLQuants) : Covariates information. In this case all are economic indicators.

multiAdv (list) : Information on TAC and quota share.

multiMainC (list) : Settings to control the main function FLBEIA.

multiBioC (list) : Settings to control the biological operating model for each stock in multiBio object.

multiFlC (list) : Settings to control the fleet operating model for each fleet in fleets object.

multiCvC (list) : Settings to control the covar operating model for each covariate in covars object. In this case all the covariates are fixed.

multiObsC (list) : Settings to control the observation model for each stock in multiBio object. In this case the stock is observed without error and there are no indices available.

multiAssC (list) : Settings to control the assessment model for each stock in multiBio object. In this case, no assessment is carried out.

multiAdvC (list) : Settings to control the advice model for each stock in multiBio object.

res_flbeia: A dataset with the outputs of FLBEIA runs given as input the different datasets (one, oneIt and multi).

oneRes (list) : Output of the FLBEIA function, given as input the data in the one dataset.

onIteRes (list) : Output of the FLBEIA function, given as input the data in the oneIt dataset.

multiRes (list) : Output of the FLBEIA function, given as input the data in the multi dataset.

mur: A dataset for Stripped Red Mullet in the Bay of Biscay. Information on catch and abundance indices from Evohe survey.

catch (data.frame) : The total catch time series data by area from WGBIE report (ICES, 2017). Total catch data is available since 1975. In 1999 France did not report any data. As France is the main contributor to the total catch, the 1999 catch data was not included in the analysis.

evhoe (data.frame) : EVHOE abundance index time series, provided by Ifremer. The abundance index is available since 1997 and provides an estimation of the biomass together with a coefficient of variation.

Datasets can be loaded by issuing the data command, like in: data(one).

All available datasets can be checked by: data(package='FLBEIA').

References

ICES, 2017

See also

Examples

data(one)
data(res_flbeia)