Extract landings, discard or catch at age

Extract landings, discards or catch at age of a stock from a FLFleetExt or FLFleetsExt object.

tlandStock(obj, stknm)

tdiscStock(obj, stknm)

landStock.f(obj, stock)

discStock.f(obj, stock)

catchStock.f(obj, stock)

landWStock.f(obj, stock)

discWStock.f(obj, stock)

catchWStock.f(obj, stock)

landStock(obj, stock)

discStock(obj, stock)

catchStock(obj, stock)

landWStock(obj, stock)

discWStock(obj, stock)

catchWStock(obj, stock)

Arguments

obj

a FLFleetExt (.f) or FLFleetsExt object

stknm

stock name

stock

stock name

Value

A FLQuant object with landings, discards or catch (total or at age).

Details

landStock.f: Extract total landings at age of a stock from a FLFleetExt obj.

discStock.f: Extract total discards at age of a stock from a FLFleetExt obj.

catchStock.f: Extract total catch at age of a stock from a FLFleetExt obj.

landWStock.f: Extract total landings at age in weight of a stock from a FLFleetExt obj.

discWStock.f: Extract total discards at age in weight of a stock from a FLFleetExt obj.

catchWStock.f: Extract total catch at age in weight of a stock from a FLFleetExt obj.

landStock: Extract total landings at age of a stock from a FLFleetExts obj.

discStock: Extract total discards at age of a stock from a FLFleetExts obj.

catchStock: Extract total catch at age of a stock from a FLFleetExts obj.

landWStock: Extract total landings at age in weight of a stock from a FLFleetExts obj.

discWStock: Extract total discards at age in weight of a stock from a FLFleetExts obj.

catchWStock: Extract total catch at age in weight of a stock from a FLFleetExts obj.

tlandStock: Total landings of a stock across fleets and metiers

tdiscStock: Total discard of a stock across fleets and metiers

Examples

## Not run: ------------------------------------
# library(FLBEIA)
# data(one)
# landWStock.f(oneFl$fl1,"stk1")
# landWStock(oneFl,"stk1")
## ---------------------------------------------