FLBEIA easy conditioning: fleets argument creator

create.fleets.data function creates an FLFleetsExt object

create.fleets.data(yrs, ns, ni, fls.data, stks.data)

Arguments

yrs
A vector with c(first.yr,proj.yr, last.yr) where

first.yr: First year of simulation (number).

proj.yr: First year of projection (number).

last.yr: Last year of projection (number).

ns
Number of seasons (number).

ni
Number of iterations (number).

fls.data
A list with the name of the fleets and the following elements:

fl.met: Name of the metiers in the fleet 'fl' (vector).

fl.met.stks: Name of the stocks in the metier 'met' and fleet 'fl' (vector).

fl_effort.flq: 'fl' fleet's effort (FLQuant).

fl.met_effshare.flq: 'fl' fleet and 'met' metier's effort share (FLQuant).

fl.met.stk_landings.n.flq: 'fl' fleet,'met' metier and 'stk' stock's landings in numbers at age.

fl_proj.avg.yrs: Historic years to calculate the average of effort, fcost, crewshare, capacity in 'fl' fleet (vector) in the projection years.

Optionals:

fl_capacity.flq: 'fl' fleet's capacity (FLQuant).

fl_fcost.flq: 'fl' fleet's fixed cost (FLQuant).

fl_crewshare.flq: 'fl' fleet's crewshare (FLQuant).

fl.met_vcost.flq: 'fl' fleet and 'met' metier's variable costs (FLQuant).

fl.met.stk_landings.wt.flq: 'fl' fleet,'met' metier and 'stk' stock's mean weight of landings at age.

fl.met.stk_discards.n.flq: 'fl' fleet,'met' metier and 'stk' stock's discards in numbers at age (FLQuant).

fl.met.stk_discards.wt.flq: 'fl' fleet,'met' metier and 'stk' stock's mean weight of discards at age.

fl.met.stk_price.flq: 'fl' fleet,'met' metier and 'stk' stock's price at age (FLQuant).

fl.met.stk_alpha.flq: 'fl' fleet,'met' metier and 'stk' stock's Cobb Douglass alpha parameter (FLQuant).

fl.met.stk_beta.flq: 'fl' fleet,'met' metier and 'stk' stock's Cobb Douglass beta parameter (FLQuant).

fl.met.stk_catch.q.flq: 'fl' fleet,'met' metier and 'stk' stock's Cobb Douglass catch.q parameter (FLQuant).

fl.met_proj.avg.yrs: Historic years to calculate the average of effshare,vcost for 'fl' fleet and 'met' metier in the projection period (vector).

fl.met.stk_proj.avg.yrs: Historic years to calculate the average of landings.wt, discards.wt, landings.sel, discards.sel alpha,beta,catch.q for 'fl' fleet, 'met' metier and 'stk' stock in the projection years (vector).

stks.data
A list with the name of the stocks and with the next elements:

stk.unit: Number of units of the stock (number).

stk.age.min: Minimum age of the stock (number).

stk.age.max: Maximum age of the stock (number).

Optionals:

stk_wt.flq: Mean weight at age of an individual (FLQuant). Required if fl.met.stk_landings.wt.flq is not defined.

stk_n.flq: Numbers at age in the population(FLQuant). Required if Cobb Douglas parameters are not defined.

stk_gB.flq: Biomass growth for the stock modeled in biomass (FLQuant). Required if Cobb Douglas parameters are not defined.

Value

An FLFleetsExt object.