Summary plots of the FLBEIA output

Summarize the results in a plot.

plotbioSum(obj, stk.nam, Blim = NA, Bpa = NA, proj.yr = NA)

plotfltSum(obj, flt.nam, proj.yr = NA)

Arguments

obj
An object with the same format as bioSum and fltSum outputs, respectively.

stk.nam
Character with the name of the stock for which summary information will be plotted. If not defined, then the first one in obj will be selected by default.

Blim
Numeric. Blim reference point for the stock (optional argument).=NA, Bpa=NA, proj.yr=NA)

Bpa
Numeric. Bpa reference point for the stock (optional argument).

proj.yr
Numeric. Year in which projection period starts (optional argument).

flt.nam
Character with the name of the fleet for which summary information will be plotted. If not defined, then the first one in obj will be selected by default.

Value

Plot.

Details

plotbioSum: Plot summarising information on stock. With one plot for each of the following indicators: "catch", "rec", "f" and "ssb". Input object should have the same format as the output of bioSum function.

plotfltSum: Plot summarising information on fleet's economic indicators. With one plot for each of the following indicators: "catch", "effort", "grossValue" and "grossSurplus". Input object should have the same format as the output of fltSum function.

Examples

# NOT RUN {
library(FLBEIA)

# Apply the summary plots to the examples runs in FLBEIA help page.

data(res_flbeia)
#------------------------------------------------
# Example One: One stock, one fleet, one iter.
#------------------------------------------------

# Biological indicators.
plotbioSum( bioSum(oneRes), Blim=800, Bpa=1200, proj.yr=2010) # bioSum output in wide format
plotbioSum( bioSum(oneRes, long = FALSE))   # bioSum output in long format
plotbioSum( bioSum(oneRes), stk.nam='stk0') # if incorrect name for the stock

# Indicators at fleet level.
plotfltSum( fltSum(oneRes), proj.yr=2010) # fltSum output in wide format
plotfltSum( fltSum(oneRes, long = FALSE)) # fltSum output in long format
plotfltSum( fltSum(oneRes), flt.nam='stk1') # if incorrect name for the fleet
plotfltSum( fltSum(oneRes, byyear = FALSE)) # although seasonal disagregation, 
                                            # it is summarised by year

#------------------------------------------------
# Example OneIters: As one but with iterations.
#------------------------------------------------

# Biological indicators.
plotbioSum( bioSum(s1), Blim=800, Bpa=1200, proj.yr=2010) # bioSum output in wide format
plotbioSum( bioSum(s1, long = FALSE))   # bioSum output in long format
plotbioSum( bioSum(s1), stk.nam='stk0') # if incorrect name for the stock

# Indicators at fleet level.
plotfltSum( fltSum(s1), proj.yr=2010) # fltSum output in wide format
plotfltSum( fltSum(s1, long = FALSE)) # fltSum output in long format
plotfltSum( fltSum(s1), flt.nam='stk1') # if incorrect name for the fleet
plotfltSum( fltSum(s1, byyear = FALSE)) # although seasonal disagregation, 
                                        # it is summarised by year

# also possible to plot information on various scenarios
sc11_bio <- bioSum(s1)
sc12_bio <- bioSum(s1, scenario='alt'); sc12_bio$value <- sc12_bio$value*1.05
plotbioSum(rbind(sc11_bio, sc12_bio), Blim=800, Bpa=1200, proj.yr=2010)

#------------------------------------------------
# Example Multi: Two stock, two fleet, four iters.
#------------------------------------------------

for (st in names(s2$stocks)) # one plot for each stock
  plotbioSum( bioSum(s2, scenario='s2'), stk.nam=st, proj.yr=2010)

for (fl in names(s2$fleets)) # one plot for each fleet
  plotfltSum( fltSum(s2, scenario='s2'), flt.nam=fl, proj.yr=2010)

# }